% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basinChars.R, R/download.R
\name{availChars}
\alias{availChars}
\title{Get Availale Basin Characteristics
The basin characteristics resource represent a list of characteristics that
defines the study area, such as Drainage Area and mean annual precipitation.}
\usage{
availChars(workspaceID, format = c("geodatabase", "shapefile"))

availChars(workspaceID, format = c("geodatabase", "shapefile"))
}
\arguments{
\item{workspaceID}{string        Service workspace received from watershed
service result}

\item{format}{either "geodatabase" or "shapefile"}

\item{rcode}{StreamStats 2-3 character code that identifies the Study Area}

\item{group}{(optional) Key word parameter group filter.}
}
\description{
Get Availale Basin Characteristics
The basin characteristics resource represent a list of characteristics that
defines the study area, such as Drainage Area and mean annual precipitation.

Download watershed resource based on workspace
returns a zip file containing the workspace contents, in either a
geodatabase or shape files
}

