% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowStats.R
\name{availFlowStats}
\alias{availFlowStats}
\title{Get available flow statistics for a region}
\usage{
availFlowStats(rcode)
}
\arguments{
\item{rcode}{2-3 character code that identifies the Study Area (either a
State or a Regional Study)
@export}
}
\description{
returns a list of flowstatistics available to be computed in the selected
region.
}
