% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combineWatersheds}
\alias{combineWatersheds}
\title{Combine watershed objects into single geojson-structured list}
\usage{
combineWatersheds(wslist, id)
}
\arguments{
\item{wslist}{list of watershed objects}

\item{id}{vector of identifiers to identify each watershed}
}
\description{
Returns a watershed object with elements of wslist combined into a single
FeatureCollection
}
\examples{
  data(westfield)
  data(pommoqusset)
  combo <- combineWatersheds(list(westfield, pommoqusset), id = c("wf", "pom"))
  leafletWatershed(combo)
}

