% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.R
\name{delineateWatershed}
\alias{delineateWatershed}
\title{Delineate watershed based on location specification}
\usage{
delineateWatershed(
  xlocation,
  ylocation,
  rcode = NULL,
  includeparameters = c("false", "true"),
  includeflowtypes = c("false", "true"),
  includefeatures = c("true", "false"),
  crs,
  simplify = NULL
)
}
\arguments{
\item{xlocation}{X location of the most downstream point of desired study area.}

\item{ylocation}{Y location of the most downstream point of desired study area.}

\item{rcode}{2-3 character code that identifies the Study Area (either a
State or a Regional Study)}

\item{includeparameters}{string    Comma separated list of region
parameters to compute. Default: true, will return all parameters for region}

\item{includeflowtypes}{string    Not yet implemented}

\item{includefeatures}{string    true    Comma separated list of features to
include in response. See Feature resource for more information.
Default: true, returns delineated basin and pourpoint}

\item{crs}{string ESPSG spatial reference code. The function sf::st_crs
might come in handy.}

\item{simplify}{boolean    Whether to simplify returned result, defaut: true}
}
\description{
Uses StreamStats API https://streamstats.usgs.gov/streamstatsservices/#
}
\examples{
\dontrun{
# Delineate a watershed and extract parameters using WGS83
ws1 <- delineateWatershed(xlocation = -72.9249, ylocation = 42.3170, crs = 4326,
includeparameters = "true")
}

}
