% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{downloadGIS}
\alias{downloadGIS}
\title{Download watershed resource based on workspace}
\usage{
downloadGIS(workspaceID, file, format = c("geodatabase", "shapefile"))
}
\arguments{
\item{workspaceID}{string        Service workspace received from watershed
service result}

\item{file}{Where to save the data, should be .zip}

\item{format}{either "geodatabase" or "shapefile"}
}
\description{
returns a zip file containing the workspace contents, in either a
geodatabase or shape files
}

