% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.R
\name{leafletWatershed}
\alias{leafletWatershed}
\title{Create a leaflet map of a watershed
Uses an object returned by delineateWatershed() and displays the watershed
boundary as well as the pour point.}
\usage{
leafletWatershed(watershed)
}
\arguments{
\item{watershed}{an object of class "watershed", as returned by
delineateWatershed()}
}
\description{
Create a leaflet map of a watershed
Uses an object returned by delineateWatershed() and displays the watershed
boundary as well as the pour point.
}

