% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowStats.R
\name{parse_region}
\alias{parse_region}
\title{Parse a region's results}
\usage{
parse_region(region)
}
\arguments{
\item{region}{a single component of a single `sstat_get()` sublist}
}
\description{
Parse a region's results
}
