% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowStats.R
\name{parse_result}
\alias{parse_result}
\title{Parse the results component of region's result list}
\usage{
parse_result(result)
}
\arguments{
\item{result}{a single component of a `sstat_get()` sub-sublist}
}
\description{
Parse the results component of region's result list
}
