% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowStats.R
\name{parse_stats}
\alias{parse_stats}
\title{Parse the results of `sstat_get()` into a tidy  data.frame}
\usage{
parse_stats(statslist, simplify = FALSE)
}
\arguments{
\item{statslist}{A list as returned by `sstat_get()`}
}
\description{
Parse the results of `sstat_get()` into a tidy  data.frame
}
