% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowStats.R
\name{parse_statsgroup}
\alias{parse_statsgroup}
\title{Parse a statisics group}
\usage{
parse_statsgroup(statsgroup, simplify = FALSE)
}
\arguments{
\item{statsgroup}{a single compoent of a list returned by `sstat_get()`}

\item{simplify}{if TRUE, combine results across regression regions using an area-weighted average.
This will also remove the `error` column.}
}
\description{
Parse a statisics group
}
