% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StreamAccumulator.R
\name{print.Rcpp_StreamAccumulator}
\alias{print.Rcpp_StreamAccumulator}
\title{Print a Streaming Accumulator}
\usage{
\method{print}{Rcpp_StreamAccumulator}(x, ...)
}
\arguments{
\item{x}{A \code{StreamAccumulator} object.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Print a Streaming Accumulator
}
