% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.R
\name{pullFeatureCollection}
\alias{pullFeatureCollection}
\title{extract part of watershed list corresponding to an individual feature collection
Returns a FeatureCollection.}
\usage{
pullFeatureCollection(ws, what = c("boundary", "pourpoint"))
}
\description{
extract part of watershed list corresponding to an individual feature collection
Returns a FeatureCollection.
}

