% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.R
\name{toSp}
\alias{toSp}
\title{Convert watershed to sp
Returns either SpatialPointsDataFrame or SpatialPolygonsDataFrame}
\usage{
toSp(watershed, what = c("boundary", "pourpoint"))
}
\arguments{
\item{watershed}{an object of class "watershed", as returned by
delineateWatershed(), or combineWatersheds()}

\item{what}{Either "boundary" or "pourpoint" describing what part of the
watershed object to write.}
}
\description{
Convert watershed to sp
Returns either SpatialPointsDataFrame or SpatialPolygonsDataFrame
}

