% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.R
\name{watershedByWorkspace}
\alias{watershedByWorkspace}
\title{Returns a watershed based on "workspace" info
Currently untested.}
\usage{
watershedByWorkspace(
  workspaceID,
  rcode,
  includeparameters = c("false", "true"),
  includeflowtypes = c("false", "true"),
  includefeatures = c("true", "false"),
  crs,
  simplify = NULL
)
}
\arguments{
\item{workspaceID}{string        Service workspace received from watershed
service result}

\item{rcode}{2-3 character code that identifies the Study Area (either a
State or a Regional Study)}

\item{includeparameters}{string    Comma separated list of region
parameters to compute. Default: true, will return all parameters for region}

\item{includeflowtypes}{string    Not yet implemented}

\item{includefeatures}{string    true    Comma separated list of features to
include in response. See Feature resource for more information.
Default: true, returns delineated basin and pourpoint}

\item{crs}{string ESPSG spatial reference code. The function sf::st_crs
might come in handy.}

\item{simplify}{boolean    Whether to simplify returned result, defaut: true
Uses StreamStats API https://streamstats.usgs.gov/streamstatsservices/#}
}
\description{
Returns a watershed based on "workspace" info
Currently untested.
}
