% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.R
\name{writeGeoJSON}
\alias{writeGeoJSON}
\title{Write part of a `watershed` object to file in geoJSON format.}
\usage{
writeGeoJSON(watershed, file, what = c("boundary", "pourpoint"))
}
\arguments{
\item{watershed}{an object of class "watershed", as returned by
delineateWatershed()}

\item{file}{a character given the file to which to write the resulting
geojson}

\item{what}{Either "boundary" or "pourpoint" describing what part of the
watershed object to write.}
}
\description{
Write part of a `watershed` object to file in geoJSON format.
}
