% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.R
\name{writeShapefile}
\alias{writeShapefile}
\title{Convert watershed to ESRI shapefile}
\usage{
writeShapefile(watershed, layer, dir = ".", what = c("boundary", "pourpoint"))
}
\arguments{
\item{watershed}{an object of class "watershed", as returned by
delineateWatershed(), or combineWatersheds()}

\item{layer}{Name of layer in resulting shapefile}

\item{dir}{Directory where shapefile files should be written.}

\item{what}{Either "boundary" or "pourpoint" describing what part of the
watershed object to write.}
}
\description{
Convert watershed to ESRI shapefile
}
