% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basinChars.R
\name{availChars}
\alias{availChars}
\title{Get Availale Basin Characteristics}
\usage{
availChars(rcode, group = NULL)
}
\arguments{
\item{rcode}{StreamStats 2-3 character code that identifies the Study Area}

\item{group}{(optional) Key word parameter group filter.}
}
\description{
The basin characteristics resource represent a list of characteristics that
defines the study area, such as Drainage Area and mean annual precipitation.
}

