% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{availFeatures}
\alias{availFeatures}
\title{List available features for a given workspace}
\usage{
availFeatures(workspaceID)
}
\arguments{
\item{workspaceID}{Service workspace received from watershed
service result}
}
\description{
returns a returns a collection of spatial feature names available for the
workspace.
}

