% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basinChars.R
\name{computeChars}
\alias{computeChars}
\title{Compute basin characteristics
Returns the computed basin characteristics based on the request configuration.}
\usage{
computeChars(rcode, workspaceID, includeparameters = "true")
}
\arguments{
\item{rcode}{2-3 character code that identifies the Study Area (either a
State or a Regional Study)}

\item{workspaceID}{Service workspace received from watershed
service result}

\item{includeparameters}{Comma separated list of region
parameters to compute. Default: true, will return all parameters for region}
}
\description{
Compute basin characteristics
Returns the computed basin characteristics based on the request configuration.
}

