% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowStats.R
\name{computeFlowStats}
\alias{computeFlowStats}
\title{Compute flow statistics
Returns the computed flow statistic values based on the request configuration.}
\usage{
computeFlowStats(rcode, workspaceID, includeparameters = "true")
}
\arguments{
\item{rcode}{2-3 character code that identifies the Study Area (either a
State or a Regional Study)}

\item{workspaceID}{Service workspace received from watershed
service result}

\item{includeflowtypes}{Comma separated list of flow types
to compute. Default: true, will return all flow types available for region}
}
\description{
Compute flow statistics
Returns the computed flow statistic values based on the request configuration.
}

