% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.R
\name{delineateWatershed}
\alias{delineateWatershed}
\title{Delineate watershed based on location specification
Uses StreamStats API http://streamstatsags.cr.usgs.gov/streamstatsservices/#}
\usage{
delineateWatershed(xlocation, ylocation, rcode = NULL,
  includeparameters = c("false", "true"), includeflowtypes = c("false",
  "true"), includefeatures = c("true", "false"), crs = NULL,
  simplify = NULL)
}
\arguments{
\item{xlocation}{X location of the most downstream point of desired study area.}

\item{ylocation}{Y location of the most downstream point of desired study area.}

\item{rcode}{2-3 character code that identifies the Study Area (either a
State or a Regional Study)}

\item{includeparameters}{string    Comma separated list of region
parameters to compute. Default: true, will return all parameters for region}

\item{includeflowtypes}{string    Not yet implemented}

\item{includefeatures}{string    true    Comma separated list of features to
include in response. See Feature resource for more information.
Default: true, returns delineated basin and pourpoint}

\item{crs}{string ESPSG spatial reference code. The function rgdal::showEPSG
might come in handy.}

\item{simplify}{boolean    Whether to simplify returned result, defaut: true}
}
\description{
Delineate watershed based on location specification
Uses StreamStats API http://streamstatsags.cr.usgs.gov/streamstatsservices/#
}

