% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamstats.R
\name{latlon2state}
\alias{latlon2state}
\title{Copied from http://stackoverflow.com/a/8751965}
\usage{
latlon2state(lat, lon)
}
\description{
library(sp)
library(maps)
library(maptools)
The single argument to this function, pointsDF, is a data.frame in which:
- column 1 contains the longitude in degrees (negative in the US)
- column 2 contains the latitude in degrees
}
